requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-api-logger==1.0.4
idna==2.10
packaging==20.9
Pillow==8.1.0
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.3
webencodings==0.5.1
